Refactored project structure to Go standards: moved config to internal/
